// src/enums/transport-mode.ts
export enum TransportMode {
  Walking = 'walking',
  Car = 'car',
  Bike = 'bike',
  BakeSharingSystem = 'bss',
}

export enum PublicTransportMode {
  Air = 'physical_mode:Air',
  Boat = 'physical_mode:Boat',
  Bus = 'physical_mode:Bus',
  BusRapidTransit = 'physical_mode:BusRapidTransit',
  Coach = 'physical_mode:Coach',
  Ferry = 'physical_mode:Ferry',
  Funicular = 'physical_mode:Funicular',
  LocalTrain = 'physical_mode:LocalTrain',
  LongDistanceTrain = 'physical_mode:LongDistanceTrain',
  Metro = 'physical_mode:Metro',
  RailShuttle = 'physical_mode:RailShuttle',
  RapidTransit = 'physical_mode:RapidTransit',
  Shuttle = 'physical_mode:Shuttle',
  Taxi = 'physical_mode:Taxi',
  Train = 'physical_mode:Train',
  Tramway = 'physical_mode:Tramway',
}